Move TypeConverter mixin to mixin.rb: parser.rb is wrong place to write it
