+++ lib/fluent/parser.rb
@@ -108,74 +108,9 @@ module Fluent
       end
     end
 
-    module TypeConverter
-      Converters = {
-        'string' => lambda { |v| v.to_s },
-        'integer' => lambda { |v| v.to_i },
-        'float' => lambda { |v| v.to_f },
-        'bool' => lambda { |v|
-          case v.downcase
-          when 'true', 'yes', '1'
-            true
-          else
-            false
-          end
-        },
-        'time' => lambda { |v, time_parser|
-          time_parser.parse(v)
-        },
-        'array' => lambda { |v, delimiter|
-          v.to_s.split(delimiter)
-        }
-      }
-
-      def self.included(klass)
-        klass.instance_eval {
-          config_param :types, :string, default: nil
-          config_param :types_delimiter, :string, default: ','
-          config_param :types_label_delimiter, :string, default: ':'
-        }
-      end
-
-      def configure(conf)
-        super
-
-        @type_converters = parse_types_parameter unless @types.nil?
-      end
-
-      private
-
-      def convert_type(name, value)
-        converter = @type_converters[name]
-        converter.nil? ? value : converter.call(value)
-      end
-
-      def parse_types_parameter
-        converters = {}
-
-        @types.split(@types_delimiter).each { |pattern_name|
-          name, type, format = pattern_name.split(@types_label_delimiter, 3)
-          raise ConfigError, "Type is needed" if type.nil?
-
-          case type
-          when 'time'
-            t_parser = TimeParser.new(format)
-            converters[name] = lambda { |v|
-              Converters[type].call(v, t_parser)
-            }
-          when 'array'
-            delimiter = format || ','
-            converters[name] = lambda { |v|
-              Converters[type].call(v, delimiter)
-            }
-          else
-            converters[name] = Converters[type]
-          end
-        }
-
-        converters
-      end
-    end
+    # TODO: will be removed at v1
+    require 'fluent/mixin'
+    TypeConverter = Fluent::TypeConverter
 
     class RegexpParser < Parser
       include TypeConverter
